Palette Id to long , more color palettes

+++ LightShow_preparation/LightShow_preparation.ino
@@ -28,7 +28,7 @@ typedef struct {
     char slot_tag; // 
     byte pattern_id;       // 0-255
     byte preset_speed_id;       // 0-255
-    byte color_palette_id;
+    long color_palette_id;
 } t_program_slot;
 
 t_program_slot g_program_slot[MAX_NUMBER_OF_PROGRAM_SLOTS]={ /* IDOL 126 BPM*/
@@ -201,7 +201,7 @@ void parse_slot_settings(String slot_setting_string)  {
   int slot_index=0;
   int pattern_id=-1;
   int preset_speed_id=-1;
-  int color_palette_id;
+  long color_palette_id;
   int slash_index;
   int colon_index;
   int token_end_index;
